Refactor: 반응형 & 스타일 수정
