+++ src/app/shopinfo/ShopRegistration.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
-import Registration from '@/components/shopinfoPage/Registartion';
+import Registration from '@/components/shopinfoPage/Registration';
 import Button from '@/components/common/Button';
 import styles from './page.module.scss';
 
 export default function ShopRegistration() {
-  const [showShopRegistartion, setShowShopRegisteration] = useState(false);
+  const [showShopRegistration, setShowShopRegisteration] = useState(false);
   return (
     <section className={styles.layout}>
       <h2 className={styles.title}>내 가게</h2>
@@ -15,7 +15,7 @@ export default function ShopRegistration() {
             가게 등록하기
           </Button>
         </div>
-        {showShopRegistartion && <Registration onClose={() => setShowShopRegisteration(false)} />}
+        {showShopRegistration && <Registration onClose={() => setShowShopRegisteration(false)} />}
       </article>
     </section>
   );
