+++ src/app/shopinfo/page.module.scss
@@ -1,9 +1,11 @@
 @import '@/styles/_index';
+@import '@/styles/_media';
 
 .layout {
   font-size: 1.6rem;
   padding: 3rem;
   flex-grow: 1;
+  max-width: 1200px;
 }
 
 .title {
@@ -25,7 +27,7 @@
   width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 2rem;
 }
 
 .mainImage {
@@ -64,16 +66,49 @@
 }
 
 .modalBox {
-  width: 100%;
   height: 100%;
+  z-index: 100;
   position: absolute;
   top: 0;
+  bottom: 0;
   left: 0;
-  z-index: 1000;
+  right: 0;
 }
 
 .cardList {
   display: grid;
+  place-items: center;
+  width: 100%;
   grid-template-columns: repeat(3, 1fr);
-  gap: 1rem;
+  grid-column-gap: 1.4rem;
+}
+
+@include media(mobile) {
+  .shopBox {
+    flex-direction: column;
+  }
+
+  .mainImage {
+    width: 100%;
+    height: auto;
+  }
+
+  .buttonBox {
+    gap: 0.6rem;
+  }
+
+  .registerLayout {
+    padding: 10rem 1rem;
+    gap: 1.6rem;
+  }
+
+  .registerButton {
+    padding: 0 1rem;
+  }
+}
+
+@include media(tablet) {
+  .cardList {
+    grid-template-columns: repeat(2, 1fr);
+  }
 }
