+++ src/components/common/Input/Input.module.scss
@@ -17,7 +17,7 @@
 
 .input {
   width: 100%;
-  border: 1px solid $gray-400;
+  border: 1px solid $gray-500;
   padding: 1.6rem 2rem;
   margin-top: 0.8rem;
