+++ src/components/common/Modal/CompletionModal/CompletionModal.module.scss
@@ -1,8 +1,8 @@
 @import '@/styles/_index';
 
 .layout {
-  width: 30rem;
-  height: 10rem;
+  width: 40rem;
+  height: 20rem;
   display: flex;
   justify-content: center;
   align-items: center;
